Adding relation between Topic and Plan

+++ src/main/java/com/github/eduriol/training/plan/app/controller/TopicRestController.java
@@ -2,13 +2,14 @@ package com.github.eduriol.training.plan.app.controller;
 
 import com.github.eduriol.training.plan.app.exception.NotFoundException;
 import com.github.eduriol.training.plan.app.exception.BadRequestException;
+import com.github.eduriol.training.plan.app.models.domain.Plan;
 import com.github.eduriol.training.plan.app.models.domain.Topic;
+import com.github.eduriol.training.plan.app.service.IPlanService;
 import com.github.eduriol.training.plan.app.service.ITopicService;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 
-import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.*;
@@ -17,7 +18,7 @@ import javax.validation.Valid;
 import java.util.List;
 
 @RestController
-@RequestMapping("/api/topics")
+@RequestMapping("/api/plans/{plan_id}/topics")
 public class TopicRestController {
 
     private static final Logger logger = LoggerFactory.getLogger(TopicRestController.class);
@@ -25,11 +26,20 @@ public class TopicRestController {
     @Autowired
     private ITopicService topicService;
 
+    @Autowired
+    private IPlanService planService;
+
     @PostMapping
     @ResponseStatus(HttpStatus.CREATED)
-    public Topic createTopic(@Valid @RequestBody Topic topic, BindingResult result) {
+    public Topic createTopic(@PathVariable Long plan_id, @Valid @RequestBody Topic topic, BindingResult result) {
 
-        logger.info("Received request to create topic: {}", topic);
+        logger.info("Received request to create topic: {} in plan with id = {}", topic, plan_id);
+
+        Plan plan = planService.findById(plan_id);
+
+        if (plan == null) {
+            throw new NotFoundException(List.of("The plan with id = " + plan_id.toString() + " does not exist."));
+        }
 
         if (result.hasErrors()) {
             List<String> errors = result.getFieldErrors()
@@ -39,6 +49,7 @@ public class TopicRestController {
             throw new BadRequestException(errors);
         }
 
+        topic.setPlan(plan);
         Topic createdTopic = topicService.save(topic);
         logger.info("Response: {}", createdTopic);
 
@@ -46,39 +57,45 @@ public class TopicRestController {
     }
 
     @GetMapping
-    public List<Topic> getAllTopics() {
-        logger.info("Received request to list all topics");
-        List<Topic> topics = topicService.findAll();
+    public List<Topic> getTopicsFromPlan(@PathVariable Long plan_id) {
+        logger.info("Received request to list all topics in plan with id = {}", plan_id);
+
+        Plan plan = planService.findById(plan_id);
+
+        if (plan == null) {
+            throw new NotFoundException(List.of("The plan with id = " + plan_id.toString() + " does not exist."));
+        }
+
+        List<Topic> topics = topicService.findByPlan(plan);
         logger.info("Response: {}", topics);
         return topics;
     }
 
     @GetMapping("/{id}")
-    public Topic getTopic(@PathVariable Long id) {
+    public Topic getTopic(@PathVariable Long id, @PathVariable Long plan_id) {
 
-        logger.info("Received request to get topic with id = {}", id);
+        logger.info("Received request to get topic with id = {} for plan with id = {}", id, plan_id);
 
-        Topic topic = topicService.findById(id);
+        Topic topic = topicService.findByIdAndPlanId(id, plan_id);
 
         if (topic == null) {
-            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " does not exist."));
+            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " and plan id = " + plan_id.toString() + " does not exist."));
         }
 
         logger.info("Response: {}", topic);
 
         return topic;
-
     }
 
     @PutMapping("/{id}")
-    public Topic updateTopic(@PathVariable Long id, @Valid @RequestBody Topic newTopic, BindingResult result) {
+    public Topic updateTopic(@PathVariable Long id, @PathVariable Long plan_id, @Valid @RequestBody Topic newTopic, BindingResult result) {
 
-        logger.info("Received request to update topic with id = {} with topic: {}", id, newTopic);
+        logger.info("Received request to update topic with id = {} and plan id = {} with topic: {}", id, plan_id, newTopic);
 
-        Topic topic = topicService.findById(id);
+        Topic topic = topicService.findByIdAndPlanId(id, plan_id);
 
         if (topic == null) {
-            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " does not exist."));
+            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " and plan id = " + plan_id.toString() + " does not exist."));
         }
 
         if (result.hasErrors()) {
@@ -99,17 +116,19 @@ public class TopicRestController {
 
     @DeleteMapping("/{id}")
     @ResponseStatus(HttpStatus.NO_CONTENT)
-    public void deleteTopic(@PathVariable Long id) {
+    public void deleteTopic(@PathVariable Long id, @PathVariable Long plan_id) {
 
-        logger.info("Received request to delete topic with id = {}", id);
+        logger.info("Received request to delete topic with id = {} in plan with id = {}", id, plan_id);
 
-        try {
-            topicService.delete(id);
-        } catch (EmptyResultDataAccessException ex) {
-            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " does not exist."));
+        Topic topic = topicService.findByIdAndPlanId(id, plan_id);
+
+        if (topic == null) {
+            throw new NotFoundException(List.of("The topic with id = " + id.toString() + " and plan id = " + plan_id.toString() + " does not exist."));
         }
 
-        logger.info("Topic with id = {} successfully deleted", id);
+        topicService.delete(id);
+
+        logger.info("Topic with id = {} in plan with id = {} successfully deleted", id, plan_id);
 
     }
